Address PR #35 review: pure cost aggregation, row-edit audit parity, citizen effort UI
- Move cost aggregation into backend/core/cost_aggregation.py as a pure,
  read-only function; persistence stays solely in compute_total_cost_for_session
  (stops the export writing model costs over live costs; removes core->router import)
- Row effort + wage-rate edits reject no-op payloads (422) and write edit_audit
  rows via the shared value_changed/insert_edit_audit_row helpers
- Validate wage_source_kind against the addressee in one shared helper used by
  both /process-steps/personnel-effort-edit and /sessions/wage-rates
- Hide the empty Reserve B/C/D effort columns for citizens (getVisibleSlots)
- Remove trailing blank lines flagged by git diff --check

Point-by-point responses to the review comments are posted in PR #35.
